build: 我没话说了

+++ build-mac-arm.sh
@@ -7,30 +7,37 @@ OUTPUT=${BASE}/output-mac-arm64-${BUILD_MODE}
 
 build-dep(){
   rm -rf build_dep && mkdir -p build_dep && cd build_dep
-  git clone https://code.videolan.org/videolan/x264.git -b stable --depth 1
-  cd x264
-  ./configure \
-    --prefix=${INSTALL_DIR} \
-    --enable-static \
-    --disable-cli \
-    --enable-pic
-  make -j$(sysctl -n hw.ncpu) install
-  cd ${BASE}
 
-  # build x265
-  rm -rf x265_build
-  mkdir x265_build && cd x265_build
-  git clone https://bitbucket.org/multicoreware/x265_git.git -b 3.6 --depth 1
-  cd x265_git/source
-  cmake -DCMAKE_INSTALL_PREFIX=${INSTALL_DIR} -DENABLE_SHARED=OFF -DCMAKE_BUILD_TYPE=Release ..
-  make -j$(sysctl -n hw.ncpu) install
+  if [[ "${BUILD_MODE}" == "gpl" ]]; then
+    git clone https://code.videolan.org/videolan/x264.git -b stable --depth 1
+    cd x264
+    ./configure \
+      --prefix=${INSTALL_DIR} \
+      --enable-static \
+      --disable-cli \
+      --enable-pic
+    make -j$(sysctl -n hw.ncpu) install
+    cd ${BASE}
+
+    # build x265
+    rm -rf x265_build
+    mkdir x265_build && cd x265_build
+    git clone https://bitbucket.org/multicoreware/x265_git.git --depth 1
+    cd x265_git
+    git fetch origin tag 3.6
+    git checkout 3.6
+    cd source
+    cmake -DCMAKE_INSTALL_PREFIX=${INSTALL_DIR} -DENABLE_SHARED=OFF -DCMAKE_BUILD_TYPE=Release ..
+    make -j$(sysctl -n hw.ncpu) install
+  fi
+
   cd ${BASE}
 }
 
 compile_ffmpeg(){
+  export PKG_CONFIG_PATH=${INSTALL_DIR}/lib/pkgconfig
   EXTRA=""
   if [[ "${BUILD_MODE}" == "gpl" ]];then
-    export PKG_CONFIG_PATH=${INSTALL_DIR}/lib/pkgconfig
     EXTRA="--enable-gpl --enable-libx264 --enable-libx265"
   fi
